+++ src/routes/api/v2/link.js
@@ -7,6 +7,9 @@ const router = express.Router();
  */
 const isDeleteKeyVaild = require('../../../middleware/api/isDeleteKeyVaild');
 const isTokenVaild = require('../../../middleware/api/isTokenVaild');
+const isQueryTokenVaild = require('../../../middleware/api/isQueryTokenVaild');
+const isBannedAPI = require('../../../middleware/api/isBanned');
+const isSuspendedAPI = require('../../../middleware/api/isSuspended');
 
 /**
  * Load Controllers
@@ -18,13 +21,20 @@ const linkController = require('../../../controllers/api/v2/link');
  */
 const linkValidation = require('../../../validation/api/v2/link');
 
-// router.get(
-//   '/',
-// );
+router.get(
+  '/',
+  isQueryTokenVaild,
+  isBannedAPI,
+  isSuspendedAPI,
+  linkController.getLinks
+);
+router.get('/:code', linkController.getLink);
 
 router.post(
   '/',
   isTokenVaild,
+  isBannedAPI,
+  isSuspendedAPI,
   linkValidation,
   linkController.createLink
 );