+++ src/routes/api/v2/upload.js
@@ -7,20 +7,24 @@ const router = express.Router();
  */
 const isTokenVaild = require('../../../middleware/api/isTokenVaild');
 const fileCheck = require('../../../middleware/api/fileCheck');
+const isBannedAPI = require('../../../middleware/api/isBanned');
+const isSuspendedAPI = require('../../../middleware/api/isSuspended');
 
 /**
  * Load Controllers
  */
-const uploadController = require('../../../controllers/api/v1/upload');
+const uploadController = require('../../../controllers/api/v2/upload');
 
 /**
  * Load Validation
  */
-const uploadValidation = require('../../../validation/api/v1/upload');
+const uploadValidation = require('../../../validation/api/v2/upload');
 
 router.post(
   '/',
   isTokenVaild,
+  isBannedAPI,
+  isSuspendedAPI,
   uploadValidation,
   fileCheck,
   uploadController.uploadFile